@@ -26,7 +26,7 @@ authors = [
 ]
 license = {text = "BSD-3-Clause"}
 urls = {Repository = "https://github.com/colmap/colmap"}
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = ["numpy"]
 classifiers = [
   "License :: OSI Approved :: BSD License",
@@ -39,10 +39,16 @@ cmake.source-dir = "python/"
 wheel.packages = ["python/pycolmap"]
 
 [tool.cibuildwheel]
-build = "cp3{9,10,11,12,13,14}-{macosx,manylinux,win}*"
+build = "cp3{10,11,12,13,14}-{macosx,manylinux,win}*"
 archs = ["auto64"]
-test-requires = "pytest mypy==1.17.0 enlighten==1.13.0"
-test-command = "python -c \"import pycolmap; print(pycolmap.__version__)\" &&  python -m mypy --package pycolmap --implicit-optional && pytest {project}/python/examples/custom_incremental_pipeline_test.py"
+test-requires = "pytest>=8.0 mypy==1.19.1 enlighten==1.13.0"
+test-command = """\
+    python -c "import pycolmap; print(pycolmap.__version__)" && \
+    cd {project} && \
+    mypy --package pycolmap && \
+    mypy --install-types --non-interactive python benchmark && \
+    pytest\
+"""
 
 [tool.cibuildwheel.linux]
 before-all = "{project}/python/ci/install-colmap-almalinux.sh"
@@ -54,3 +60,19 @@ before-all = "{project}/python/ci/install-colmap-macos.sh"
 before-all = "pwsh -File {project}/python/ci/install-colmap-windows.ps1"
 before-build = "pip install delvewheel"
 test-command = "pwsh -File {project}/python/ci/test-colmap-windows.ps1"
+
+[tool.pytest.ini_options]
+minversion = "8.0"
+addopts = ["-ra", "-q"]
+testpaths = [
+    "python",
+    "benchmark",
+]
+norecursedirs = [
+    "build",
+    "python/build",
+]
+
+[tool.mypy]
+implicit_optional = true
+ignore_missing_imports = true
