@@ -5,7 +5,6 @@ Python reimplementation of the C++ incremental mapper with equivalent logic.
 import argparse
 import time
 from pathlib import Path
-from typing import Optional
 
 import custom_bundle_adjustment
 import enlighten
@@ -22,7 +21,7 @@ def write_snapshot(
     path = snapshot_path / f"{timestamp:010d}"
     path.mkdir(exist_ok=True, parents=True)
     logging.verbose(1, f"=> Writing to {path}")
-    reconstruction.write(path)
+    reconstruction.write(str(path))
 
 
 def iterative_global_refinement(
@@ -65,13 +64,13 @@ def initialize_reconstruction(
         if not all(reconstruction.exists_image(i) for i in init_pair):
             logging.info(f"=> Initial image pair {init_pair} does not exist.")
             return pycolmap.IncrementalMapperStatus.BAD_INITIAL_PAIR
-        init_cam2_from_cam1 = mapper.estimate_initial_two_view_geometry(
+        maybe_init_cam2_from_cam1 = mapper.estimate_initial_two_view_geometry(
             mapper_options, *init_pair
         )
-        if init_cam2_from_cam1 is None:
+        if maybe_init_cam2_from_cam1 is None:
             logging.info("Provided pair is insuitable for initialization")
             return pycolmap.IncrementalMapperStatus.BAD_INITIAL_PAIR
-
+        init_cam2_from_cam1 = maybe_init_cam2_from_cam1
     logging.info(
         f"Registering initial image pair #{init_pair[0]} and #{init_pair[1]}"
     )
@@ -79,11 +78,10 @@ def initialize_reconstruction(
         mapper_options, *init_pair, init_cam2_from_cam1
     )
     for image_id in init_pair:
-        for data_id in reconstruction.images[image_id].frame.data_ids:
-            if data_id.sensor_id.type == pycolmap.SensorType.CAMERA:
-                mapper.triangulate_image(
-                    options.get_triangulation(), data_id.id
-                )
+        image = reconstruction.images[image_id]
+        assert image.frame is not None
+        for data_id in image.frame.image_ids:
+            mapper.triangulate_image(options.get_triangulation(), data_id.id)
 
     logging.info("Global bundle adjustment")
     # The following is equivalent to: mapper.adjust_global_bundle(...)
@@ -199,11 +197,12 @@ def reconstruct_sub_model(
             if reg_next_success:
                 break
         if reg_next_success:
-            for data_id in reconstruction.images[next_image_id].frame.data_ids:
-                if data_id.sensor_id.type == pycolmap.SensorType.CAMERA:
-                    mapper.triangulate_image(
-                        options.get_triangulation(), data_id.id
-                    )
+            image = reconstruction.images[next_image_id]
+            assert image.frame is not None
+            for data_id in image.frame.image_ids:
+                mapper.triangulate_image(
+                    options.get_triangulation(), data_id.id
+                )
             # This is equivalent to mapper.iterative_local_refinement(...)
             custom_bundle_adjustment.iterative_local_refinement(
                 mapper,
@@ -331,7 +330,10 @@ def main_incremental_mapper(controller: pycolmap.IncrementalPipeline) -> None:
     """Equivalent to IncrementalPipeline.run()"""
     timer = pycolmap.Timer()
     timer.start()
-    if not controller.load_database():
+
+    database_cache = controller.database_cache
+    if database_cache.num_images() == 0:
+        logging.warning("No images with matches found in the database")
         return
 
     reconstruction_manager = controller.reconstruction_manager
@@ -343,7 +345,6 @@ def main_incremental_mapper(controller: pycolmap.IncrementalPipeline) -> None:
             "but multiple are given"
         )
 
-    database_cache = controller.database_cache
     mapper = pycolmap.IncrementalMapper(database_cache)
     mapper_options = controller.options.get_mapper()
     reconstruct(controller, mapper, mapper_options, continue_reconstruction)
@@ -377,8 +378,8 @@ def main(
     database_path: Path,
     image_path: Path,
     output_path: Path,
-    options: Optional[pycolmap.IncrementalPipelineOptions] = None,
-    input_path: Optional[Path] = None,
+    options: pycolmap.IncrementalPipelineOptions | None = None,
+    input_path: Path | None = None,
 ) -> dict[int, pycolmap.Reconstruction]:
     if options is None:
         options = pycolmap.IncrementalPipelineOptions()
@@ -389,10 +390,10 @@ def main(
         logging.fatal(f"Image path does not exist: {image_path}")
     output_path.mkdir(exist_ok=True, parents=True)
     reconstruction_manager = pycolmap.ReconstructionManager()
-    if input_path is not None and input_path != "":
-        reconstruction_manager.read(input_path)
+    if input_path:
+        reconstruction_manager.read(str(input_path))
 
-    with pycolmap.Database.open(database_path) as database:
+    with pycolmap.Database.open(str(database_path)) as database:
         mapper = pycolmap.IncrementalPipeline(
             options, database, reconstruction_manager
         )
@@ -413,7 +414,7 @@ def main(
                 main_incremental_mapper(mapper)
 
     # write and output
-    reconstruction_manager.write(output_path)
+    reconstruction_manager.write(str(output_path))
     reconstructions = {}
     for i in range(reconstruction_manager.size()):
         reconstructions[i] = reconstruction_manager.get(i)
