@@ -63,7 +63,9 @@ except ImportError as e:
 if TYPE_CHECKING:
     from ._core import *  # noqa F403
 
-__all__ = import_module_symbols(globals(), _core, exclude={"cost_functions"})
+__all__ = import_module_symbols(
+    globals(), _core, exclude={"cost_functions", "pyceres"}
+)
 __all__.extend(["__version__", "__ceres_version__"])
 
 __version__ = _core.__version__
