@@ -44,7 +44,7 @@ class BundleAdjustmentController : public BaseController {
   void Run();
 
  private:
-  const OptionManager options_;
+  const OptionManager& options_;
   std::shared_ptr<Reconstruction> reconstruction_;
 };
 
