@@ -86,8 +86,9 @@ TEST(GlobalPipeline, SfMWithRandomSeedStability) {
   synthetic_noise_options.point2D_stddev = 0.5;
   SynthesizeNoise(synthetic_noise_options, &gt_reconstruction, database.get());
 
-  auto run_mapper = [&](int random_seed) {
+  auto run_mapper = [&](int num_threads, int random_seed) {
     glomap::GlobalMapperOptions options;
+    options.num_threads = num_threads;
     options.random_seed = random_seed;
     auto reconstruction_manager = std::make_shared<ReconstructionManager>();
     GlobalPipeline mapper(options, database, reconstruction_manager);
@@ -98,17 +99,32 @@ TEST(GlobalPipeline, SfMWithRandomSeedStability) {
 
   constexpr int kRandomSeed = 42;
 
-  // Running with the same seed should produce similar results.
-  // Due to multi-threading, we allow small floating-point variations.
-  auto reconstruction_manager0 = run_mapper(kRandomSeed);
-  auto reconstruction_manager1 = run_mapper(kRandomSeed);
-  EXPECT_THAT(*reconstruction_manager0->Get(0),
-              ReconstructionNear(*reconstruction_manager1->Get(0),
-                                 /*max_rotation_error_deg=*/1e-10,
-                                 /*max_proj_center_error=*/1e-10,
-                                 /*max_scale_error=*/std::nullopt,
-                                 /*num_obs_tolerance=*/0.01,
-                                 /*align=*/false));
+  // Single-threaded execution.
+  {
+    auto reconstruction_manager0 =
+        run_mapper(/*num_threads=*/1, /*random_seed=*/kRandomSeed);
+    auto reconstruction_manager1 =
+        run_mapper(/*num_threads=*/1, /*random_seed=*/kRandomSeed);
+    EXPECT_THAT(*reconstruction_manager0->Get(0),
+                ReconstructionEq(*reconstruction_manager1->Get(0)));
+  }
+
+  // Multi-threaded execution.
+  {
+    auto reconstruction_manager0 =
+        run_mapper(/*num_threads=*/3, /*random_seed=*/kRandomSeed);
+    auto reconstruction_manager1 =
+        run_mapper(/*num_threads=*/3, /*random_seed=*/kRandomSeed);
+    // Same seed should produce similar results, up to floating-point variations
+    // in optimization.
+    EXPECT_THAT(*reconstruction_manager0->Get(0),
+                ReconstructionNear(*reconstruction_manager1->Get(0),
+                                   /*max_rotation_error_deg=*/1e-9,
+                                   /*max_proj_center_error=*/1e-9,
+                                   /*max_scale_error=*/std::nullopt,
+                                   /*num_obs_tolerance=*/0.01,
+                                   /*align=*/false));
+  }
 }
 
 }  // namespace
