@@ -32,8 +32,7 @@
 #include "colmap/util/logging.h"
 #include "colmap/util/timer.h"
 
-#include "glomap/io/colmap_io.h"
-#include "glomap/processors/view_graph_manipulation.h"
+#include "glomap/sfm/global_mapper.h"
 
 namespace colmap {
 
@@ -46,64 +45,44 @@ RotationAveragingController::RotationAveragingController(
       reconstruction_(std::move(THROW_CHECK_NOTNULL(reconstruction))) {}
 
 void RotationAveragingController::Run() {
-  // Initialize view graph from database.
-  glomap::ViewGraph view_graph;
-  glomap::InitializeGlomapFromDatabase(
-      *database_, *reconstruction_, view_graph);
-
-  // Read pose priors from database.
-  std::vector<PosePrior> pose_priors = database_->ReadAllPosePriors();
-
-  if (view_graph.Empty()) {
-    LOG(ERROR) << "Cannot continue without image pairs";
-    return;
+  // Propagate options to component options.
+  RotationAveragingControllerOptions options = options_;
+  if (options.random_seed >= 0) {
+    options.rotation_estimation.random_seed = options.random_seed;
+    options.view_graph_calibration.random_seed = options.random_seed;
   }
+  options.view_graph_calibration.solver_options.num_threads =
+      options.num_threads;
 
   Timer run_timer;
   run_timer.Start();
 
-  // Step 0: Preprocessing
-  LOG(INFO) << "----- Running preprocessing -----";
-  glomap::ViewGraphManipulator::UpdateImagePairsConfig(view_graph,
-                                                       *reconstruction_);
-  glomap::ViewGraphManipulator::DecomposeRelPose(view_graph, *reconstruction_);
+  // Create a global mapper instance
+  glomap::GlobalMapper mapper(database_);
+  mapper.BeginReconstruction(reconstruction_);
 
-  // Step 1: View graph calibration
+  if (mapper.ViewGraph()->Empty()) {
+    LOG(ERROR) << "Cannot continue without image pairs";
+    return;
+  }
+
   LOG(INFO) << "----- Running view graph calibration -----";
-  glomap::ViewGraphCalibrator calibrator(options_.view_graph_calibration);
-  if (!calibrator.Solve(view_graph, *reconstruction_)) {
+  if (!glomap::CalibrateViewGraph(options.view_graph_calibration,
+                                  *mapper.ViewGraph(),
+                                  *reconstruction_)) {
     LOG(ERROR) << "Failed to solve view graph calibration";
     return;
   }
 
-  // Step 2: Relative pose estimation
-  LOG(INFO) << "----- Running relative pose estimation -----";
-  glomap::EstimateRelativePoses(
-      view_graph, *reconstruction_, options_.relative_pose_estimation);
-
-  glomap::ImagePairsInlierCount(
-      view_graph, *reconstruction_, options_.inlier_thresholds, true);
-
-  view_graph.FilterByNumInliers(options_.inlier_thresholds.min_inlier_num);
-  view_graph.FilterByInlierRatio(options_.inlier_thresholds.min_inlier_ratio);
-
-  if (view_graph.KeepLargestConnectedComponents(*reconstruction_) == 0) {
-    LOG(ERROR) << "No connected components found";
-    return;
-  }
-
-  // Step 3: Rotation averaging
   LOG(INFO) << "----- Running rotation averaging -----";
-  if (!glomap::SolveRotationAveraging(options_.rotation_estimation,
-                                      view_graph,
-                                      *reconstruction_,
-                                      pose_priors)) {
+  if (!mapper.RotationAveraging(options.rotation_estimation,
+                                options.max_rotation_error_deg)) {
     LOG(ERROR) << "Failed to solve rotation averaging";
     return;
   }
 
   LOG(INFO) << "Rotation averaging done in " << run_timer.ElapsedSeconds()
-            << " seconds";
+            << "s";
 }
 
 }  // namespace colmap
