@@ -627,7 +627,12 @@ TwoViewGeometry EstimateCalibratedTwoViewGeometry(
 
   // Estimate epipolar models.
 
-  auto E_ransac_options = options.ransac_options;
+  auto ransac_options = options.ransac_options;
+  if (options.min_inlier_ratio > 0) {
+    ransac_options.min_inlier_ratio = options.min_inlier_ratio;
+  }
+
+  auto E_ransac_options = ransac_options;
   E_ransac_options.max_error =
       (camera1.CamFromImgThreshold(options.ransac_options.max_error) +
        camera2.CamFromImgThreshold(options.ransac_options.max_error)) /
@@ -640,7 +645,7 @@ TwoViewGeometry EstimateCalibratedTwoViewGeometry(
 
   LORANSAC<FundamentalMatrixSevenPointEstimator,
            FundamentalMatrixEightPointEstimator>
-      F_ransac(options.ransac_options);
+      F_ransac(ransac_options);
   const auto F_report =
       F_ransac.Estimate(matched_img_points1, matched_img_points2);
   geometry.F = F_report.model;
@@ -648,7 +653,7 @@ TwoViewGeometry EstimateCalibratedTwoViewGeometry(
   // Estimate planar or panoramic model.
 
   LORANSAC<HomographyMatrixEstimator, HomographyMatrixEstimator> H_ransac(
-      options.ransac_options);
+      ransac_options);
   const auto H_report =
       H_ransac.Estimate(matched_img_points1, matched_img_points2);
   geometry.H = H_report.model;
@@ -727,6 +732,16 @@ TwoViewGeometry EstimateCalibratedTwoViewGeometry(
   if (best_inlier_mask != nullptr) {
     geometry.inlier_matches =
         ExtractInlierMatches(matches, num_inliers, *best_inlier_mask);
+
+    // Check inlier ratio threshold.
+    if (options.min_inlier_ratio > 0) {
+      const double inlier_ratio =
+          static_cast<double>(num_inliers) / matches.size();
+      if (inlier_ratio < options.min_inlier_ratio) {
+        geometry.config = TwoViewGeometry::ConfigurationType::DEGENERATE;
+        return geometry;
+      }
+    }
 
     if (options.detect_watermark && DetectWatermarkMatches(camera1,
                                                            matched_img_points1,
