@@ -63,10 +63,9 @@ class Database {
   Database() = default;
 
   // Closes the database, if not closed before.
-  virtual ~Database();
+  virtual ~Database() = 0;
 
-  NON_COPYABLE(Database);  // Use Clone() instead.
-  NON_MOVABLE(Database);
+  NON_COPYABLE(Database);
 
   // Factory function to create a database implementation for a given path.
   // The factory should be robust to handle non-supported files and return a
@@ -84,9 +83,6 @@ class Database {
 
   // Explicitly close the database before destruction.
   virtual void Close() = 0;
-
-  // Return a clone of the database.
-  virtual std::shared_ptr<Database> Clone() const = 0;
 
   // Check if entry already exists in database. For image pairs, the order of
   // `image_id1` and `image_id2` does not matter.
