@@ -63,14 +63,21 @@ class DatabaseCache {
   //                              same frame will also be included. All images
   //                              are used if empty.
   void Load(const Database& database,
-            size_t min_num_matches,
-            bool ignore_watermarks,
-            const std::unordered_set<std::string>& image_names);
+            size_t min_num_matches = 0,
+            bool ignore_watermarks = false,
+            const std::unordered_set<std::string>& image_names = {});
 
   static std::shared_ptr<DatabaseCache> Create(
       const Database& database,
+      size_t min_num_matches = 0,
+      bool ignore_watermarks = false,
+      const std::unordered_set<std::string>& image_names = {});
+
+  // Create a filtered database cache from an existing cache containing only
+  // the specified images and their associated data.
+  static std::shared_ptr<DatabaseCache> CreateFromCache(
+      const DatabaseCache& database_cache,
       size_t min_num_matches,
-      bool ignore_watermarks,
       const std::unordered_set<std::string>& image_names);
 
   // Get number of objects.
