@@ -1,5 +1,6 @@
 #include "glomap/controllers/option_manager.h"
 
+#include "glomap/estimators/global_positioning.h"
 #include "glomap/estimators/gravity_refinement.h"
 #include "glomap/sfm/global_mapper.h"
 
@@ -24,17 +25,12 @@ void OptionManager::AddGlobalMapperOptions() {
   added_global_mapper_options_ = true;
 
   // Global mapper options
+  AddAndRegisterDefaultOption("Mapper.num_threads", &mapper->num_threads);
   AddAndRegisterDefaultOption("Mapper.random_seed", &mapper->random_seed);
   AddAndRegisterDefaultOption("Mapper.num_iterations_ba",
                               &mapper->num_iterations_ba);
-  AddAndRegisterDefaultOption("Mapper.num_iterations_retriangulation",
-                              &mapper->num_iterations_retriangulation);
-  AddAndRegisterDefaultOption("Mapper.skip_preprocessing",
-                              &mapper->skip_preprocessing);
   AddAndRegisterDefaultOption("Mapper.skip_view_graph_calibration",
                               &mapper->skip_view_graph_calibration);
-  AddAndRegisterDefaultOption("Mapper.skip_relative_pose_estimation",
-                              &mapper->skip_relative_pose_estimation);
   AddAndRegisterDefaultOption("Mapper.skip_rotation_averaging",
                               &mapper->skip_rotation_averaging);
   AddAndRegisterDefaultOption("Mapper.skip_global_positioning",
@@ -47,19 +43,14 @@ void OptionManager::AddGlobalMapperOptions() {
 
   // View graph calibration options
   AddAndRegisterDefaultOption(
-      "ViewGraphCalib.thres_lower_ratio",
-      &mapper->view_graph_calibration.thres_lower_ratio);
+      "ViewGraphCalib.min_focal_length_ratio",
+      &mapper->view_graph_calibration.min_focal_length_ratio);
   AddAndRegisterDefaultOption(
-      "ViewGraphCalib.thres_higher_ratio",
-      &mapper->view_graph_calibration.thres_higher_ratio);
+      "ViewGraphCalib.max_focal_length_ratio",
+      &mapper->view_graph_calibration.max_focal_length_ratio);
   AddAndRegisterDefaultOption(
-      "ViewGraphCalib.thres_two_view_error",
-      &mapper->view_graph_calibration.thres_two_view_error);
-
-  // Relative pose estimation options
-  AddAndRegisterDefaultOption(
-      "RelPoseEstimation.max_epipolar_error",
-      &mapper->relative_pose_estimation.ransac_options.max_epipolar_error);
+      "ViewGraphCalib.max_calibration_error",
+      &mapper->view_graph_calibration.max_calibration_error);
 
   // Track establishment options
   AddAndRegisterDefaultOption(
@@ -90,6 +81,13 @@ void OptionManager::AddGlobalMapperOptions() {
   AddAndRegisterDefaultOption(
       "GlobalPositioning.max_num_iterations",
       &mapper->global_positioning.solver_options.max_num_iterations);
+  AddAndRegisterDefaultEnumOption(
+      "GlobalPositioning.constraint_type",
+      &mapper->global_positioning.constraint_type,
+      GlobalPositioningConstraintTypeToString,
+      GlobalPositioningConstraintTypeFromString,
+      "{ONLY_POINTS, ONLY_CAMERAS, POINTS_AND_CAMERAS_BALANCED, "
+      "POINTS_AND_CAMERAS}");
 
   // Bundle adjustment options
   AddAndRegisterDefaultOption("BundleAdjustment.use_gpu",
@@ -115,39 +113,24 @@ void OptionManager::AddGlobalMapperOptions() {
       "BundleAdjustment.max_num_iterations",
       &mapper->bundle_adjustment.solver_options.max_num_iterations);
 
-  // Triangulation options
+  // Retriangulation options
   AddAndRegisterDefaultOption(
-      "Triangulation.complete_max_reproj_error",
-      &mapper->retriangulation.tri_complete_max_reproj_error);
-  AddAndRegisterDefaultOption(
-      "Triangulation.merge_max_reproj_error",
-      &mapper->retriangulation.tri_merge_max_reproj_error);
-  AddAndRegisterDefaultOption("Triangulation.min_angle",
-                              &mapper->retriangulation.tri_min_angle);
-  AddAndRegisterDefaultOption("Triangulation.min_num_matches",
-                              &mapper->retriangulation.min_num_matches);
+      "Retriangulation.complete_max_reproj_error",
+      &mapper->retriangulation.complete_max_reproj_error);
+  AddAndRegisterDefaultOption("Retriangulation.merge_max_reproj_error",
+                              &mapper->retriangulation.merge_max_reproj_error);
+  AddAndRegisterDefaultOption("Retriangulation.min_angle",
+                              &mapper->retriangulation.min_angle);
 
-  // Inlier threshold options
-  AddAndRegisterDefaultOption("Thresholds.max_angle_error",
-                              &mapper->inlier_thresholds.max_angle_error);
-  AddAndRegisterDefaultOption(
-      "Thresholds.max_reprojection_error",
-      &mapper->inlier_thresholds.max_reprojection_error);
-  AddAndRegisterDefaultOption(
-      "Thresholds.min_triangulation_angle",
-      &mapper->inlier_thresholds.min_triangulation_angle);
-  AddAndRegisterDefaultOption("Thresholds.max_epipolar_error_E",
-                              &mapper->inlier_thresholds.max_epipolar_error_E);
-  AddAndRegisterDefaultOption("Thresholds.max_epipolar_error_F",
-                              &mapper->inlier_thresholds.max_epipolar_error_F);
-  AddAndRegisterDefaultOption("Thresholds.max_epipolar_error_H",
-                              &mapper->inlier_thresholds.max_epipolar_error_H);
-  AddAndRegisterDefaultOption("Thresholds.min_inlier_num",
-                              &mapper->inlier_thresholds.min_inlier_num);
-  AddAndRegisterDefaultOption("Thresholds.min_inlier_ratio",
-                              &mapper->inlier_thresholds.min_inlier_ratio);
-  AddAndRegisterDefaultOption("Thresholds.max_rotation_error",
-                              &mapper->inlier_thresholds.max_rotation_error);
+  // Threshold options
+  AddAndRegisterDefaultOption("Mapper.max_rotation_error_deg",
+                              &mapper->max_rotation_error_deg);
+  AddAndRegisterDefaultOption("Mapper.max_angular_reproj_error_deg",
+                              &mapper->max_angular_reproj_error_deg);
+  AddAndRegisterDefaultOption("Mapper.max_normalized_reproj_error",
+                              &mapper->max_normalized_reproj_error);
+  AddAndRegisterDefaultOption("Mapper.min_tri_angle_deg",
+                              &mapper->min_tri_angle_deg);
 }
 
 void OptionManager::AddGravityRefinerOptions() {
@@ -170,8 +153,6 @@ void OptionManager::Reset() {
   added_global_mapper_options_ = false;
   added_gravity_refiner_options_ = false;
 }
-
-bool OptionManager::Check() { return colmap::BaseOptionManager::Check(); }
 
 void OptionManager::ResetOptions(const bool reset_paths) {
   colmap::BaseOptionManager::ResetOptions(reset_paths);
