@@ -2,7 +2,6 @@
 
 #include "colmap/geometry/pose.h"
 #include "colmap/math/spanning_tree.h"
-#include "colmap/scene/reconstruction_io_utils.h"
 
 #include "glomap/estimators/rotation_averaging_impl.h"
 
@@ -62,7 +61,7 @@ image_t ComputeMaximumSpanningTree(
       continue;
     }
     edges.emplace_back(it1->second, it2->second);
-    weights.push_back(static_cast<float>(image_pair.inliers.size()));
+    weights.push_back(static_cast<float>(image_pair.inlier_matches.size()));
   }
 
   // Compute spanning tree using generic algorithm.
@@ -155,9 +154,28 @@ bool RotationEstimator::MaybeSolveGravityAlignedSubset(
 
   if (should_solve) {
     LOG(INFO) << "Solving subset 1-DOF rotation averaging problem";
+    colmap::Reconstruction gravity_reconstruction(reconstruction);
+    gravity_view_graph.KeepLargestConnectedComponents(gravity_reconstruction);
     if (!SolveRotationAveraging(
-            gravity_view_graph, pose_priors, reconstruction)) {
+            gravity_view_graph, pose_priors, gravity_reconstruction)) {
       return false;
+    }
+
+    for (const auto& [gravity_frame_id, gravity_frame] :
+         gravity_reconstruction.Frames()) {
+      if (!gravity_frame.HasPose()) continue;
+      reconstruction.Frame(gravity_frame_id)
+          .SetRigFromWorld(gravity_frame.RigFromWorld());
+    }
+
+    for (const auto& [gravity_rig_id, gravity_rig] :
+         gravity_reconstruction.Rigs()) {
+      for (const auto& [sensor_id, sensor_from_rig] :
+           gravity_rig.NonRefSensors()) {
+        if (!gravity_rig.HasSensorFromRig(sensor_id)) continue;
+        reconstruction.Rig(gravity_rig_id)
+            .SetSensorFromRig(sensor_id, sensor_from_rig);
+      }
     }
   }
 
@@ -473,7 +491,7 @@ colmap::Reconstruction CreateExpandedReconstruction(
 bool SolveRotationAveraging(const RotationEstimatorOptions& options,
                             ViewGraph& view_graph,
                             colmap::Reconstruction& reconstruction,
-                            std::vector<colmap::PosePrior>& pose_priors) {
+                            const std::vector<colmap::PosePrior>& pose_priors) {
   view_graph.KeepLargestConnectedComponents(reconstruction);
 
   // If there are cameras with unknown cam_from_rig, run expanded rotation
