@@ -1,7 +1,6 @@
 #pragma once
 
 #include "colmap/scene/reconstruction.h"
-#include "colmap/sensor/models.h"
 
 #include "glomap/scene/view_graph.h"
 
@@ -12,15 +11,24 @@
 namespace glomap {
 
 struct ViewGraphCalibratorOptions {
-  // The minimal ratio of the estimated focal length against the prior focal
-  // length
-  double thres_lower_ratio = 0.1;
-  // The maximal ratio of the estimated focal length against the prior focal
-  // length
-  double thres_higher_ratio = 10;
+  // Random seed for RANSAC-based estimation (-1 for random).
+  int random_seed = -1;
 
-  // The threshold for the corresponding error in the problem for an image pair
-  double thres_two_view_error = 2.;
+  // Whether to cross-validate prior focal lengths by checking the ratio of
+  // calibrated vs uncalibrated pairs per camera. When enabled, UNCALIBRATED
+  // pairs are converted to CALIBRATED if both cameras have reliable priors.
+  bool cross_validate_prior_focal_lengths = true;
+
+  // Whether to re-estimate relative poses after focal length calibration.
+  bool reestimate_relative_pose = true;
+
+  // The minimum ratio of the estimated focal length to the prior focal length.
+  double min_focal_length_ratio = 0.1;
+  // The maximum ratio of the estimated focal length to the prior focal length.
+  double max_focal_length_ratio = 10;
+
+  // The maximum calibration error for an image pair.
+  double max_calibration_error = 2.;
 
   // Scaling factor for the loss function
   double loss_function_scale = 0.01;
@@ -28,52 +36,30 @@ struct ViewGraphCalibratorOptions {
   // The options for the solver
   ceres::Solver::Options solver_options;
 
+  // Options for relative pose re-estimation.
+  double relpose_max_error = 1.0;
+  int relpose_min_num_inliers = 30;
+  double relpose_min_inlier_ratio = 0.25;
+
   ViewGraphCalibratorOptions() {
     solver_options.num_threads = -1;
     solver_options.max_num_iterations = 100;
     solver_options.function_tolerance = 1e-5;
   }
 
-  std::shared_ptr<ceres::LossFunction> CreateLossFunction() {
-    return std::make_shared<ceres::CauchyLoss>(loss_function_scale);
+  // Create loss function for given options.
+  std::unique_ptr<ceres::LossFunction> CreateLossFunction() const {
+    return std::make_unique<ceres::CauchyLoss>(loss_function_scale);
   }
 };
 
-class ViewGraphCalibrator {
- public:
-  explicit ViewGraphCalibrator(const ViewGraphCalibratorOptions& options)
-      : options_(options) {}
-
-  // Entry point for the calibration
-  bool Solve(ViewGraph& view_graph, colmap::Reconstruction& reconstruction);
-
- private:
-  // Reset the problem
-  void Reset(const colmap::Reconstruction& reconstruction);
-
-  // Add the image pairs to the problem
-  void AddImagePairsToProblem(const ViewGraph& view_graph,
-                              const colmap::Reconstruction& reconstruction);
-
-  // Add a single image pair to the problem
-  void AddImagePair(image_t image_id1,
-                    image_t image_id2,
-                    const ImagePair& image_pair,
-                    const colmap::Reconstruction& reconstruction);
-
-  // Set the cameras to be constant if they have prior intrinsics
-  size_t ParameterizeCameras(const colmap::Reconstruction& reconstruction);
-
-  // Convert the results back to the camera
-  void CopyBackResults(colmap::Reconstruction& reconstruction);
-
-  // Filter the image pairs based on the calibration results
-  size_t FilterImagePairs(ViewGraph& view_graph) const;
-
-  ViewGraphCalibratorOptions options_;
-  std::unique_ptr<ceres::Problem> problem_;
-  std::unordered_map<camera_t, double> focals_;
-  std::shared_ptr<ceres::LossFunction> loss_function_;
-};
+// Calibrate the view graph by estimating focal lengths from fundamental
+// matrices. Filters image pairs with high calibration errors.
+// Then re-estimates relative poses using the calibrated cameras.
+// See: "Stable Intrinsic Auto-Calibration from Fundamental Matrices of Devices
+// with Uncorrelated Camera Parameters", Fetzer et al., WACV 2020.
+bool CalibrateViewGraph(const ViewGraphCalibratorOptions& options,
+                        ViewGraph& view_graph,
+                        colmap::Reconstruction& reconstruction);
 
 }  // namespace glomap
