@@ -30,6 +30,7 @@
 #include "glomap/scene/view_graph.h"
 
 #include "colmap/scene/synthetic.h"
+#include "colmap/util/testing.h"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
@@ -47,9 +48,9 @@ ImagePair SynthesizeImagePair(int num_inliers = 50, int num_matches = 100) {
     pair.matches(i, 1) = i;
   }
   // First num_inliers matches are inliers.
-  pair.inliers.resize(num_inliers);
+  pair.inlier_matches.reserve(num_inliers);
   for (int i = 0; i < num_inliers; ++i) {
-    pair.inliers[i] = i;
+    pair.inlier_matches.emplace_back(i, i);
   }
   return pair;
 }
@@ -257,48 +258,6 @@ TEST(ViewGraph, ValidImagePairs) {
               testing::UnorderedElementsAre(pair_id1, pair_id2, pair_id3));
 }
 
-TEST(ViewGraph, FilterByNumInliers) {
-  ViewGraph view_graph;
-
-  const image_pair_t pair_id1 = colmap::ImagePairToPairId(1, 2);
-  const image_pair_t pair_id2 = colmap::ImagePairToPairId(1, 3);
-  const image_pair_t pair_id3 = colmap::ImagePairToPairId(2, 3);
-  const image_pair_t pair_id4 = colmap::ImagePairToPairId(2, 4);
-  view_graph.AddImagePair(1, 2, SynthesizeImagePair(50));
-  view_graph.AddImagePair(1, 3, SynthesizeImagePair(20));
-  view_graph.AddImagePair(2, 3, SynthesizeImagePair(30));
-  view_graph.AddImagePair(2, 4, SynthesizeImagePair(50));
-  view_graph.SetInvalidImagePair(pair_id4);  // Already invalid
-
-  view_graph.FilterByNumInliers(30);
-
-  EXPECT_TRUE(view_graph.IsValid(pair_id1));
-  EXPECT_FALSE(view_graph.IsValid(pair_id2));
-  EXPECT_TRUE(view_graph.IsValid(pair_id3));
-  EXPECT_FALSE(view_graph.IsValid(pair_id4));
-}
-
-TEST(ViewGraph, FilterByInlierRatio) {
-  ViewGraph view_graph;
-
-  const image_pair_t pair_id1 = colmap::ImagePairToPairId(1, 2);
-  const image_pair_t pair_id2 = colmap::ImagePairToPairId(1, 3);
-  const image_pair_t pair_id3 = colmap::ImagePairToPairId(2, 3);
-  const image_pair_t pair_id4 = colmap::ImagePairToPairId(2, 4);
-  view_graph.AddImagePair(1, 2, SynthesizeImagePair(50));  // 50% ratio
-  view_graph.AddImagePair(1, 3, SynthesizeImagePair(10));  // 10% ratio
-  view_graph.AddImagePair(2, 3, SynthesizeImagePair(25));  // 25% ratio
-  view_graph.AddImagePair(2, 4, SynthesizeImagePair(50));
-  view_graph.SetInvalidImagePair(pair_id4);  // Already invalid
-
-  view_graph.FilterByInlierRatio(0.25);
-
-  EXPECT_TRUE(view_graph.IsValid(pair_id1));
-  EXPECT_FALSE(view_graph.IsValid(pair_id2));
-  EXPECT_TRUE(view_graph.IsValid(pair_id3));
-  EXPECT_FALSE(view_graph.IsValid(pair_id4));
-}
-
 TEST(ViewGraph, FilterByRelativeRotation) {
   colmap::Reconstruction reconstruction;
   colmap::SyntheticDatasetOptions options;
@@ -346,6 +305,65 @@ TEST(ViewGraph, FilterByRelativeRotation) {
   EXPECT_FALSE(view_graph.IsValid(pair_id2));
   EXPECT_TRUE(view_graph.IsValid(pair_id3));
   EXPECT_FALSE(view_graph.IsValid(pair_id4));
+}
+
+TEST(ViewGraph, LoadFromDatabase) {
+  const std::string test_dir = colmap::CreateTestDir();
+
+  // Create two databases with overlapping image pairs.
+  auto database1 = colmap::Database::Open(test_dir + "/database1.db");
+  auto database2 = colmap::Database::Open(test_dir + "/database2.db");
+
+  colmap::Camera camera = colmap::Camera::CreateFromModelId(
+      colmap::kInvalidCameraId,
+      colmap::SimplePinholeCameraModel::model_id,
+      1,
+      1,
+      1);
+  const colmap::camera_t camera_id1 = database1->WriteCamera(camera);
+  const colmap::camera_t camera_id2 = database2->WriteCamera(camera);
+
+  // Create images in both databases.
+  for (int i = 1; i <= 4; ++i) {
+    colmap::Image image;
+    image.SetName("image" + std::to_string(i));
+    image.SetCameraId(camera_id1);
+    database1->WriteImage(image);
+    image.SetCameraId(camera_id2);
+    database2->WriteImage(image);
+  }
+
+  colmap::TwoViewGeometry two_view;
+  two_view.config = colmap::TwoViewGeometry::CALIBRATED;
+  two_view.inlier_matches = {{0, 0}, {1, 1}};
+
+  // Database1: pairs (1,2) and (2,3)
+  database1->WriteMatches(1, 2, colmap::FeatureMatches(10));
+  database1->WriteMatches(2, 3, colmap::FeatureMatches(10));
+  database1->WriteTwoViewGeometry(1, 2, two_view);
+  database1->WriteTwoViewGeometry(2, 3, two_view);
+
+  // Database2: pairs (2,3) and (3,4) - pair (2,3) overlaps with database1
+  database2->WriteMatches(2, 3, colmap::FeatureMatches(10));
+  database2->WriteMatches(3, 4, colmap::FeatureMatches(10));
+  database2->WriteTwoViewGeometry(2, 3, two_view);
+  database2->WriteTwoViewGeometry(3, 4, two_view);
+
+  ViewGraph view_graph;
+
+  // First read from database1 should succeed.
+  view_graph.LoadFromDatabase(*database1);
+  EXPECT_EQ(view_graph.NumImagePairs(), 2);
+
+  // Second read from database2 with allow_duplicate=false should throw
+  // because pair (2,3) already exists.
+  EXPECT_ANY_THROW(
+      view_graph.LoadFromDatabase(*database2, /*allow_duplicate=*/false));
+
+  // Second read from database2 with allow_duplicate=true should succeed.
+  view_graph.LoadFromDatabase(*database2, /*allow_duplicate=*/true);
+  // Should now have 3 pairs: (1,2), (2,3), (3,4)
+  EXPECT_EQ(view_graph.NumImagePairs(), 3);
 }
 
 }  // namespace
