@@ -1,351 +1,437 @@
 #include "glomap/sfm/global_mapper.h"
 
+#include "colmap/scene/database_cache.h"
 #include "colmap/scene/projection.h"
+#include "colmap/sfm/incremental_mapper.h"
 #include "colmap/sfm/observation_manager.h"
 #include "colmap/util/logging.h"
 #include "colmap/util/timer.h"
 
 #include "glomap/estimators/rotation_averaging.h"
-#include "glomap/processors/image_pair_inliers.h"
 #include "glomap/processors/reconstruction_pruning.h"
 #include "glomap/processors/view_graph_manipulation.h"
 
 namespace glomap {
 
-// TODO: Rig normalizaiton has not be done
-bool GlobalMapper::Solve(const colmap::Database* database,
-                         ViewGraph& view_graph,
-                         colmap::Reconstruction& reconstruction,
-                         std::vector<colmap::PosePrior>& pose_priors,
+GlobalMapper::GlobalMapper(std::shared_ptr<const colmap::Database> database) {
+  THROW_CHECK_NOTNULL(database);
+  // TODO: Directly use DatabaseCache in the signature and make min_num_matches
+  // an option in the global pipeline.
+  constexpr int kMinNumMatches = 15;
+  database_cache_ = colmap::DatabaseCache::Create(*database,
+                                                  kMinNumMatches,
+                                                  /*ignore_watermarks=*/false,
+                                                  /*image_names=*/{});
+  // TODO: Move to BeginReconstruction after migrating to PoseGraph and accept
+  // DatabaseCache.
+  view_graph_ = std::make_shared<class ViewGraph>();
+  view_graph_->LoadFromDatabase(*database);
+}
+
+void GlobalMapper::BeginReconstruction(
+    const std::shared_ptr<colmap::Reconstruction>& reconstruction) {
+  THROW_CHECK_NOTNULL(reconstruction);
+  reconstruction_ = reconstruction;
+  reconstruction_->Load(*database_cache_);
+}
+
+std::shared_ptr<colmap::Reconstruction> GlobalMapper::Reconstruction() const {
+  return reconstruction_;
+}
+
+std::shared_ptr<class ViewGraph> GlobalMapper::ViewGraph() const {
+  return view_graph_;
+}
+
+bool GlobalMapper::RotationAveraging(const RotationEstimatorOptions& options,
+                                     double max_rotation_error_deg) {
+  // TODO: This is a misuse of frame registration. Frames should only be
+  // registered when their poses are actually computed, not with arbitrary
+  // identity poses. The rotation averaging code should be updated to work
+  // with unregistered frames.
+  // Register all frames with an initial identity pose so rotation averaging
+  // can work. The actual rotations will be computed by rotation averaging.
+  for (const auto& [frame_id, frame] : reconstruction_->Frames()) {
+    if (!frame.HasPose()) {
+      reconstruction_->Frame(frame_id).SetRigFromWorld(Rigid3d());
+      reconstruction_->RegisterFrame(frame_id);
+    }
+  }
+
+  if (view_graph_->KeepLargestConnectedComponents(*reconstruction_) == 0) {
+    LOG(ERROR) << "no connected components are found";
+    return false;
+  }
+
+  // Read pose priors from the database cache.
+  const std::vector<colmap::PosePrior>& pose_priors =
+      database_cache_->PosePriors();
+
+  // The first run is for filtering
+  SolveRotationAveraging(options, *view_graph_, *reconstruction_, pose_priors);
+
+  view_graph_->FilterByRelativeRotation(*reconstruction_,
+                                        max_rotation_error_deg);
+  if (view_graph_->KeepLargestConnectedComponents(*reconstruction_) == 0) {
+    LOG(ERROR) << "no connected components are found";
+    return false;
+  }
+
+  // The second run is for final estimation
+  if (!SolveRotationAveraging(
+          options, *view_graph_, *reconstruction_, pose_priors)) {
+    return false;
+  }
+  view_graph_->FilterByRelativeRotation(*reconstruction_,
+                                        max_rotation_error_deg);
+  image_t num_img =
+      view_graph_->KeepLargestConnectedComponents(*reconstruction_);
+  if (num_img == 0) {
+    LOG(ERROR) << "no connected components are found";
+    return false;
+  }
+  LOG(INFO) << num_img << " / " << reconstruction_->NumImages()
+            << " images are within the connected component.";
+
+  return true;
+}
+
+void GlobalMapper::EstablishTracks(const TrackEstablishmentOptions& options) {
+  // TrackEngine reads images, writes unfiltered points3D to a temporary map,
+  // then filters into the main reconstruction
+  std::unordered_map<point3D_t, Point3D> unfiltered_points3D;
+  TrackEngine track_engine(*view_graph_, reconstruction_->Images(), options);
+  track_engine.EstablishFullTracks(unfiltered_points3D);
+
+  // Filter the points3D into a selected subset
+  std::unordered_map<point3D_t, Point3D> selected_points3D;
+  point3D_t num_points3D =
+      track_engine.FindTracksForProblem(unfiltered_points3D, selected_points3D);
+
+  // Add selected points3D to reconstruction
+  THROW_CHECK_EQ(reconstruction_->NumPoints3D(), 0);
+  for (auto& [point3D_id, point3D] : selected_points3D) {
+    reconstruction_->AddPoint3D(point3D_id, std::move(point3D));
+  }
+  LOG(INFO) << "Before filtering: " << unfiltered_points3D.size()
+            << ", after filtering: " << num_points3D;
+}
+
+bool GlobalMapper::GlobalPositioning(const GlobalPositionerOptions& options,
+                                     double max_angular_reproj_error_deg,
+                                     double max_normalized_reproj_error,
+                                     double min_tri_angle_deg) {
+  if (options.constraint_type != GlobalPositioningConstraintType::ONLY_POINTS) {
+    LOG(ERROR) << "Only points are used for solving camera positions";
+    return false;
+  }
+
+  GlobalPositioner gp_engine(options);
+
+  // TODO: consider to support other modes as well
+  if (!gp_engine.Solve(*view_graph_, *reconstruction_)) {
+    return false;
+  }
+
+  // Filter tracks based on the estimation
+  colmap::ObservationManager obs_manager(*reconstruction_);
+
+  // First pass: use relaxed threshold (2x) for cameras without prior focal.
+  obs_manager.FilterPoints3DWithLargeReprojectionError(
+      2.0 * max_angular_reproj_error_deg,
+      reconstruction_->Point3DIds(),
+      colmap::ReprojectionErrorType::ANGULAR);
+
+  // Second pass: apply strict threshold for cameras with prior focal length.
+  const double max_angular_error_rad =
+      colmap::DegToRad(max_angular_reproj_error_deg);
+  std::vector<std::pair<colmap::image_t, colmap::point2D_t>> obs_to_delete;
+  for (const auto point3D_id : reconstruction_->Point3DIds()) {
+    if (!reconstruction_->ExistsPoint3D(point3D_id)) {
+      continue;
+    }
+    const auto& point3D = reconstruction_->Point3D(point3D_id);
+    for (const auto& track_el : point3D.track.Elements()) {
+      const auto& image = reconstruction_->Image(track_el.image_id);
+      const auto& camera = *image.CameraPtr();
+      if (!camera.has_prior_focal_length) {
+        continue;
+      }
+      const auto& point2D = image.Point2D(track_el.point2D_idx);
+      const double error = colmap::CalculateAngularReprojectionError(
+          point2D.xy, point3D.xyz, image.CamFromWorld(), camera);
+      if (error > max_angular_error_rad) {
+        obs_to_delete.emplace_back(track_el.image_id, track_el.point2D_idx);
+      }
+    }
+  }
+  for (const auto& [image_id, point2D_idx] : obs_to_delete) {
+    if (reconstruction_->Image(image_id).Point2D(point2D_idx).HasPoint3D()) {
+      obs_manager.DeleteObservation(image_id, point2D_idx);
+    }
+  }
+
+  // Filter tracks based on triangulation angle and reprojection error
+  obs_manager.FilterPoints3DWithSmallTriangulationAngle(
+      min_tri_angle_deg, reconstruction_->Point3DIds());
+  // Set the threshold to be larger to avoid removing too many tracks
+  obs_manager.FilterPoints3DWithLargeReprojectionError(
+      10 * max_normalized_reproj_error,
+      reconstruction_->Point3DIds(),
+      colmap::ReprojectionErrorType::NORMALIZED);
+
+  // Normalize the structure
+  // If the camera rig is used, the structure do not need to be normalized
+  reconstruction_->Normalize();
+
+  return true;
+}
+
+bool GlobalMapper::IterativeBundleAdjustment(
+    const BundleAdjusterOptions& options,
+    double max_normalized_reproj_error,
+    double min_tri_angle_deg,
+    int num_iterations) {
+  for (int ite = 0; ite < num_iterations; ite++) {
+    // First stage: optimize positions only (rotation constant)
+    if (!RunBundleAdjustment(options,
+                             /*constant_rotation=*/true,
+                             *reconstruction_)) {
+      return false;
+    }
+    LOG(INFO) << "Global bundle adjustment iteration " << ite + 1 << " / "
+              << num_iterations << ", stage 1 finished (position only)";
+
+    // Second stage: optimize rotations if desired
+    if (options.optimize_rotations &&
+        !RunBundleAdjustment(options,
+                             /*constant_rotation=*/false,
+                             *reconstruction_)) {
+      return false;
+    }
+    LOG(INFO) << "Global bundle adjustment iteration " << ite + 1 << " / "
+              << num_iterations << ", stage 2 finished";
+
+    // Normalize the structure
+    reconstruction_->Normalize();
+
+    // Filter tracks based on the estimation
+    // For the filtering, in each round, the criteria for outlier is
+    // tightened. If only few tracks are changed, no need to start bundle
+    // adjustment right away. Instead, use a more strict criteria to filter
+    LOG(INFO) << "Filtering tracks by reprojection ...";
+
+    colmap::ObservationManager obs_manager(*reconstruction_);
+    bool status = true;
+    size_t filtered_num = 0;
+    while (status && ite < num_iterations) {
+      double scaling = std::max(3 - ite, 1);
+      filtered_num += obs_manager.FilterPoints3DWithLargeReprojectionError(
+          scaling * max_normalized_reproj_error,
+          reconstruction_->Point3DIds(),
+          colmap::ReprojectionErrorType::NORMALIZED);
+
+      if (filtered_num > 1e-3 * reconstruction_->NumPoints3D()) {
+        status = false;
+      } else {
+        ite++;
+      }
+    }
+    if (status) {
+      LOG(INFO) << "fewer than 0.1% tracks are filtered, stop the iteration.";
+      break;
+    }
+  }
+
+  // Filter tracks based on the estimation
+  LOG(INFO) << "Filtering tracks by reprojection ...";
+  {
+    colmap::ObservationManager obs_manager(*reconstruction_);
+    obs_manager.FilterPoints3DWithLargeReprojectionError(
+        max_normalized_reproj_error,
+        reconstruction_->Point3DIds(),
+        colmap::ReprojectionErrorType::NORMALIZED);
+    obs_manager.FilterPoints3DWithSmallTriangulationAngle(
+        min_tri_angle_deg, reconstruction_->Point3DIds());
+  }
+
+  return true;
+}
+
+bool GlobalMapper::IterativeRetriangulateAndRefine(
+    const colmap::IncrementalTriangulator::Options& options,
+    const BundleAdjusterOptions& ba_options,
+    double max_normalized_reproj_error,
+    double min_tri_angle_deg) {
+  // Delete all existing 3D points and re-establish 2D-3D correspondences.
+  reconstruction_->DeleteAllPoints2DAndPoints3D();
+
+  // Initialize mapper.
+  colmap::IncrementalMapper mapper(database_cache_);
+  mapper.BeginReconstruction(reconstruction_);
+
+  // Triangulate all registered images.
+  for (const auto image_id : reconstruction_->RegImageIds()) {
+    mapper.TriangulateImage(options, image_id);
+  }
+
+  // Set up bundle adjustment options for colmap's incremental mapper.
+  colmap::BundleAdjustmentOptions colmap_ba_options;
+  colmap_ba_options.solver_options.num_threads =
+      ba_options.solver_options.num_threads;
+  colmap_ba_options.solver_options.max_num_iterations = 50;
+  colmap_ba_options.solver_options.max_linear_solver_iterations = 100;
+  colmap_ba_options.print_summary = false;
+
+  // Iterative global refinement.
+  colmap::IncrementalMapper::Options mapper_options;
+  mapper_options.random_seed = options.random_seed;
+  mapper.IterativeGlobalRefinement(/*max_num_refinements=*/5,
+                                   /*max_refinement_change=*/0.0005,
+                                   mapper_options,
+                                   colmap_ba_options,
+                                   options,
+                                   /*normalize_reconstruction=*/true);
+
+  mapper.EndReconstruction(/*discard=*/false);
+
+  // Final filtering and bundle adjustment.
+  colmap::ObservationManager obs_manager(*reconstruction_);
+  obs_manager.FilterPoints3DWithLargeReprojectionError(
+      max_normalized_reproj_error,
+      reconstruction_->Point3DIds(),
+      colmap::ReprojectionErrorType::NORMALIZED);
+
+  if (!RunBundleAdjustment(ba_options,
+                           /*constant_rotation=*/false,
+                           *reconstruction_)) {
+    return false;
+  }
+
+  reconstruction_->Normalize();
+
+  obs_manager.FilterPoints3DWithLargeReprojectionError(
+      max_normalized_reproj_error,
+      reconstruction_->Point3DIds(),
+      colmap::ReprojectionErrorType::NORMALIZED);
+  obs_manager.FilterPoints3DWithSmallTriangulationAngle(
+      min_tri_angle_deg, reconstruction_->Point3DIds());
+
+  return true;
+}
+
+// TODO: Rig normalizaiton has not been done
+bool GlobalMapper::Solve(const GlobalMapperOptions& options,
                          std::unordered_map<frame_t, int>& cluster_ids) {
-  // Propagate random seed to component options for deterministic behavior.
-  GlobalMapperOptions options = options_;
-  if (options.random_seed >= 0) {
-    options.relative_pose_estimation.random_seed = options.random_seed;
-    options.rotation_averaging.random_seed = options.random_seed;
-    options.global_positioning.random_seed = options.random_seed;
+  THROW_CHECK_NOTNULL(reconstruction_);
+  THROW_CHECK_NOTNULL(view_graph_);
+
+  // Propagate random seed and num_threads to component options.
+  GlobalMapperOptions opts = options;
+  if (opts.random_seed >= 0) {
+    opts.rotation_averaging.random_seed = opts.random_seed;
+    opts.global_positioning.random_seed = opts.random_seed;
+    opts.global_positioning.use_parameter_block_ordering = false;
+    opts.retriangulation.random_seed = opts.random_seed;
+    opts.view_graph_calibration.random_seed = opts.random_seed;
   }
+  opts.view_graph_calibration.solver_options.num_threads = opts.num_threads;
+  opts.global_positioning.solver_options.num_threads = opts.num_threads;
+  opts.bundle_adjustment.solver_options.num_threads = opts.num_threads;
 
-  // 0. Preprocessing
-  if (!options.skip_preprocessing) {
-    LOG(INFO) << "----- Running preprocessing -----";
-
-    colmap::Timer run_timer;
-    run_timer.Start();
-    // If camera intrinsics seem to be good, force the pair to use essential
-    // matrix
-    ViewGraphManipulator::UpdateImagePairsConfig(view_graph, reconstruction);
-    ViewGraphManipulator::DecomposeRelPose(view_graph, reconstruction);
-    run_timer.PrintSeconds();
-  }
-
-  // 1. Run view graph calibration
-  if (!options.skip_view_graph_calibration) {
+  if (!opts.skip_view_graph_calibration) {
+    // Run view graph calibration
     LOG(INFO) << "----- Running view graph calibration -----";
-    ViewGraphCalibrator vgcalib_engine(options.view_graph_calibration);
-    if (!vgcalib_engine.Solve(view_graph, reconstruction)) {
-      return false;
-    }
-  }
-
-  // 2. Run relative pose estimation
-  //   TODO: Use generalized relative pose estimation for rigs.
-  if (!options.skip_relative_pose_estimation) {
-    LOG(INFO) << "----- Running relative pose estimation -----";
-
     colmap::Timer run_timer;
     run_timer.Start();
-    // Relative pose relies on the undistorted images
-    EstimateRelativePoses(
-        view_graph, reconstruction, options.relative_pose_estimation);
-
-    InlierThresholdOptions inlier_thresholds = options.inlier_thresholds;
-    // Undistort the images and filter edges by inlier number
-    ImagePairsInlierCount(view_graph, reconstruction, inlier_thresholds, true);
-
-    view_graph.FilterByNumInliers(options.inlier_thresholds.min_inlier_num);
-    view_graph.FilterByInlierRatio(options.inlier_thresholds.min_inlier_ratio);
-
-    if (view_graph.KeepLargestConnectedComponents(reconstruction) == 0) {
-      LOG(ERROR) << "no connected components are found";
+    if (!CalibrateViewGraph(
+            opts.view_graph_calibration, *view_graph_, *reconstruction_)) {
       return false;
     }
-
-    run_timer.PrintSeconds();
+    LOG(INFO) << "View graph calibration done in " << run_timer.ElapsedSeconds()
+              << " seconds";
+  } else {
+    // Decompose relative poses when skipping view graph calibration.
+    ViewGraphManipulator::DecomposeRelativePoses(
+        *view_graph_, *reconstruction_, opts.num_threads);
   }
 
-  // 3. Run rotation averaging for three times
-  if (!options.skip_rotation_averaging) {
+  // Run rotation averaging
+  if (!opts.skip_rotation_averaging) {
     LOG(INFO) << "----- Running rotation averaging -----";
-
     colmap::Timer run_timer;
     run_timer.Start();
-
-    // The first run is for filtering
-    SolveRotationAveraging(
-        options.rotation_averaging, view_graph, reconstruction, pose_priors);
-
-    view_graph.FilterByRelativeRotation(
-        reconstruction, options.inlier_thresholds.max_rotation_error);
-    if (view_graph.KeepLargestConnectedComponents(reconstruction) == 0) {
-      LOG(ERROR) << "no connected components are found";
+    if (!RotationAveraging(opts.rotation_averaging,
+                           opts.max_rotation_error_deg)) {
       return false;
     }
-
-    // The second run is for final estimation
-    if (!SolveRotationAveraging(options.rotation_averaging,
-                                view_graph,
-                                reconstruction,
-                                pose_priors)) {
-      return false;
-    }
-    view_graph.FilterByRelativeRotation(
-        reconstruction, options.inlier_thresholds.max_rotation_error);
-    image_t num_img = view_graph.KeepLargestConnectedComponents(reconstruction);
-    if (num_img == 0) {
-      LOG(ERROR) << "no connected components are found";
-      return false;
-    }
-    LOG(INFO) << num_img << " / " << reconstruction.NumImages()
-              << " images are within the connected component.";
-
-    run_timer.PrintSeconds();
+    LOG(INFO) << "Rotation averaging done in " << run_timer.ElapsedSeconds()
+              << " seconds";
   }
 
-  // 4. Track establishment and selection
-  if (!options.skip_track_establishment) {
+  // Track establishment and selection
+  if (!opts.skip_track_establishment) {
     LOG(INFO) << "----- Running track establishment -----";
     colmap::Timer run_timer;
     run_timer.Start();
-
-    // TrackEngine reads images, writes unfiltered points3D to a temporary map,
-    // then filters into the main reconstruction
-    std::unordered_map<point3D_t, Point3D> unfiltered_points3D;
-    TrackEngine track_engine(
-        view_graph, reconstruction.Images(), options.track_establishment);
-    track_engine.EstablishFullTracks(unfiltered_points3D);
-
-    // Filter the points3D into a selected subset
-    std::unordered_map<point3D_t, Point3D> selected_points3D;
-    point3D_t num_points3D = track_engine.FindTracksForProblem(
-        unfiltered_points3D, selected_points3D);
-
-    // Add selected points3D to reconstruction
-    THROW_CHECK_EQ(reconstruction.NumPoints3D(), 0);
-    for (auto& [point3D_id, point3D] : selected_points3D) {
-      reconstruction.AddPoint3D(point3D_id, std::move(point3D));
-    }
-    LOG(INFO) << "Before filtering: " << unfiltered_points3D.size()
-              << ", after filtering: " << num_points3D;
-    run_timer.PrintSeconds();
+    EstablishTracks(opts.track_establishment);
+    LOG(INFO) << "Track establishment done in " << run_timer.ElapsedSeconds()
+              << " seconds";
   }
 
-  // 5. Global positioning
-  if (!options.skip_global_positioning) {
+  // Global positioning
+  if (!opts.skip_global_positioning) {
     LOG(INFO) << "----- Running global positioning -----";
-
-    if (options.global_positioning.constraint_type !=
-        GlobalPositionerOptions::ConstraintType::ONLY_POINTS) {
-      LOG(ERROR) << "Only points are used for solving camera positions";
-      return false;
-    }
-
     colmap::Timer run_timer;
     run_timer.Start();
-
-    GlobalPositioner gp_engine(options.global_positioning);
-
-    // TODO: consider to support other modes as well
-    if (!gp_engine.Solve(view_graph, reconstruction)) {
+    if (!GlobalPositioning(opts.global_positioning,
+                           opts.max_angular_reproj_error_deg,
+                           opts.max_normalized_reproj_error,
+                           opts.min_tri_angle_deg)) {
       return false;
     }
-    // Filter tracks based on the estimation
-    colmap::ObservationManager obs_manager(reconstruction);
-    // First pass: use relaxed threshold (2x) for cameras without prior focal.
-    obs_manager.FilterPoints3DWithLargeReprojectionError(
-        2.0 * options.inlier_thresholds.max_angle_error,
-        reconstruction.Point3DIds(),
-        colmap::ReprojectionErrorType::ANGULAR);
-    // Second pass: apply strict threshold for cameras with prior focal length.
-    const double max_angle_error_rad =
-        colmap::DegToRad(options.inlier_thresholds.max_angle_error);
-    std::vector<std::pair<colmap::image_t, colmap::point2D_t>> obs_to_delete;
-    for (const auto point3D_id : reconstruction.Point3DIds()) {
-      if (!reconstruction.ExistsPoint3D(point3D_id)) {
-        continue;
-      }
-      const auto& point3D = reconstruction.Point3D(point3D_id);
-      for (const auto& track_el : point3D.track.Elements()) {
-        const auto& image = reconstruction.Image(track_el.image_id);
-        const auto& camera = *image.CameraPtr();
-        if (!camera.has_prior_focal_length) {
-          continue;
-        }
-        const auto& point2D = image.Point2D(track_el.point2D_idx);
-        const double error = colmap::CalculateAngularReprojectionError(
-            point2D.xy, point3D.xyz, image.CamFromWorld(), camera);
-        if (error > max_angle_error_rad) {
-          obs_to_delete.emplace_back(track_el.image_id, track_el.point2D_idx);
-        }
-      }
-    }
-    for (const auto& [image_id, point2D_idx] : obs_to_delete) {
-      if (reconstruction.Image(image_id).Point2D(point2D_idx).HasPoint3D()) {
-        obs_manager.DeleteObservation(image_id, point2D_idx);
-      }
-    }
-
-    // Filter tracks based on triangulation angle and reprojection error
-    obs_manager.FilterPoints3DWithSmallTriangulationAngle(
-        options.inlier_thresholds.min_triangulation_angle,
-        reconstruction.Point3DIds());
-    // Set the threshold to be larger to avoid removing too many tracks
-    obs_manager.FilterPoints3DWithLargeReprojectionError(
-        10 * options.inlier_thresholds.max_reprojection_error,
-        reconstruction.Point3DIds(),
-        colmap::ReprojectionErrorType::NORMALIZED);
-    // Normalize the structure
-    // If the camera rig is used, the structure do not need to be normalized
-    reconstruction.Normalize();
-
-    run_timer.PrintSeconds();
+    LOG(INFO) << "Global positioning done in " << run_timer.ElapsedSeconds()
+              << " seconds";
   }
 
-  // 6. Bundle adjustment
-  if (!options.skip_bundle_adjustment) {
-    LOG(INFO) << "----- Running bundle adjustment -----";
-
+  // Bundle adjustment
+  if (!opts.skip_bundle_adjustment) {
+    LOG(INFO) << "----- Running iterative bundle adjustment -----";
     colmap::Timer run_timer;
     run_timer.Start();
-
-    for (int ite = 0; ite < options.num_iterations_ba; ite++) {
-      // 6.1. First stage: optimize positions only (rotation constant)
-      if (!RunBundleAdjustment(options.bundle_adjustment,
-                               /*constant_rotation=*/true,
-                               reconstruction)) {
-        return false;
-      }
-      LOG(INFO) << "Global bundle adjustment iteration " << ite + 1 << " / "
-                << options.num_iterations_ba
-                << ", stage 1 finished (position only)";
-      run_timer.PrintSeconds();
-
-      // 6.2. Second stage: optimize rotations if desired
-      if (options.bundle_adjustment.optimize_rotations &&
-          !RunBundleAdjustment(options.bundle_adjustment,
-                               /*constant_rotation=*/false,
-                               reconstruction)) {
-        return false;
-      }
-      LOG(INFO) << "Global bundle adjustment iteration " << ite + 1 << " / "
-                << options.num_iterations_ba << ", stage 2 finished";
-      if (ite != options.num_iterations_ba - 1) run_timer.PrintSeconds();
-
-      // Normalize the structure
-      reconstruction.Normalize();
-
-      // 6.3. Filter tracks based on the estimation
-      // For the filtering, in each round, the criteria for outlier is
-      // tightened. If only few tracks are changed, no need to start bundle
-      // adjustment right away. Instead, use a more strict criteria to filter
-      LOG(INFO) << "Filtering tracks by reprojection ...";
-
-      colmap::ObservationManager obs_manager(reconstruction);
-      bool status = true;
-      size_t filtered_num = 0;
-      while (status && ite < options.num_iterations_ba) {
-        double scaling = std::max(3 - ite, 1);
-        filtered_num += obs_manager.FilterPoints3DWithLargeReprojectionError(
-            scaling * options.inlier_thresholds.max_reprojection_error,
-            reconstruction.Point3DIds(),
-            colmap::ReprojectionErrorType::NORMALIZED);
-
-        if (filtered_num > 1e-3 * reconstruction.NumPoints3D()) {
-          status = false;
-        } else
-          ite++;
-      }
-      if (status) {
-        LOG(INFO) << "fewer than 0.1% tracks are filtered, stop the iteration.";
-        break;
-      }
+    if (!IterativeBundleAdjustment(opts.bundle_adjustment,
+                                   opts.max_normalized_reproj_error,
+                                   opts.min_tri_angle_deg,
+                                   opts.num_iterations_ba)) {
+      return false;
     }
-
-    // Filter tracks based on the estimation
-    LOG(INFO) << "Filtering tracks by reprojection ...";
-    {
-      colmap::ObservationManager obs_manager(reconstruction);
-      obs_manager.FilterPoints3DWithLargeReprojectionError(
-          options.inlier_thresholds.max_reprojection_error,
-          reconstruction.Point3DIds(),
-          colmap::ReprojectionErrorType::NORMALIZED);
-      obs_manager.FilterPoints3DWithSmallTriangulationAngle(
-          options.inlier_thresholds.min_triangulation_angle,
-          reconstruction.Point3DIds());
-    }
-
-    run_timer.PrintSeconds();
+    LOG(INFO) << "Iterative bundle adjustment done in "
+              << run_timer.ElapsedSeconds() << " seconds";
   }
 
-  // 7. Retriangulation
-  if (!options.skip_retriangulation) {
-    THROW_CHECK_NOTNULL(database);
-    LOG(INFO) << "----- Running retriangulation -----";
-    for (int ite = 0; ite < options.num_iterations_retriangulation; ite++) {
-      colmap::Timer run_timer;
-      run_timer.Start();
-      RetriangulateTracks(options.retriangulation, *database, reconstruction);
-      run_timer.PrintSeconds();
-
-      LOG(INFO) << "Running bundle adjustment...";
-      if (!RunBundleAdjustment(options.bundle_adjustment,
-                               /*constant_rotation=*/false,
-                               reconstruction)) {
-        return false;
-      }
-
-      // Filter tracks based on the estimation
-      LOG(INFO) << "Filtering tracks by reprojection ...";
-      colmap::ObservationManager(reconstruction)
-          .FilterPoints3DWithLargeReprojectionError(
-              options.inlier_thresholds.max_reprojection_error,
-              reconstruction.Point3DIds(),
-              colmap::ReprojectionErrorType::NORMALIZED);
-
-      // Run BA again after filtering outliers
-      if (!RunBundleAdjustment(options.bundle_adjustment,
-                               /*constant_rotation=*/false,
-                               reconstruction)) {
-        return false;
-      }
-      run_timer.PrintSeconds();
+  // Retriangulation
+  if (!opts.skip_retriangulation) {
+    LOG(INFO) << "----- Running iterative retriangulation and refinement -----";
+    colmap::Timer run_timer;
+    run_timer.Start();
+    if (!IterativeRetriangulateAndRefine(opts.retriangulation,
+                                         opts.bundle_adjustment,
+                                         opts.max_normalized_reproj_error,
+                                         opts.min_tri_angle_deg)) {
+      return false;
     }
-
-    // Normalize the structure
-    reconstruction.Normalize();
-
-    // Filter tracks based on the estimation
-    LOG(INFO) << "Filtering tracks by reprojection ...";
-    {
-      colmap::ObservationManager obs_manager(reconstruction);
-      obs_manager.FilterPoints3DWithLargeReprojectionError(
-          options.inlier_thresholds.max_reprojection_error,
-          reconstruction.Point3DIds(),
-          colmap::ReprojectionErrorType::NORMALIZED);
-      obs_manager.FilterPoints3DWithSmallTriangulationAngle(
-          options.inlier_thresholds.min_triangulation_angle,
-          reconstruction.Point3DIds());
-    }
+    LOG(INFO) << "Iterative retriangulation and refinement done in "
+              << run_timer.ElapsedSeconds() << " seconds";
   }
 
-  // 8. Reconstruction pruning
-  if (!options.skip_pruning) {
+  // Reconstruction pruning
+  if (!opts.skip_pruning) {
     LOG(INFO) << "----- Running postprocessing -----";
-
     colmap::Timer run_timer;
     run_timer.Start();
-    cluster_ids = PruneWeaklyConnectedFrames(reconstruction);
-    run_timer.PrintSeconds();
+    cluster_ids = PruneWeaklyConnectedFrames(*reconstruction_);
+    LOG(INFO) << "Postprocessing done in " << run_timer.ElapsedSeconds()
+              << " seconds";
   }
 
   return true;
