@@ -4,8 +4,6 @@
 #include "colmap/scene/synthetic.h"
 #include "colmap/util/testing.h"
 
-#include "glomap/io/colmap_io.h"
-
 #include <gtest/gtest.h>
 
 namespace glomap {
@@ -16,7 +14,6 @@ namespace {
 GlobalMapperOptions CreateTestOptions() {
   GlobalMapperOptions options;
   options.skip_view_graph_calibration = false;
-  options.skip_relative_pose_estimation = false;
   options.skip_rotation_averaging = false;
   options.skip_track_establishment = false;
   options.skip_global_positioning = false;
@@ -38,19 +35,16 @@ TEST(GlobalMapper, WithoutNoise) {
   colmap::SynthesizeDataset(
       synthetic_dataset_options, &gt_reconstruction, database.get());
 
-  ViewGraph view_graph;
-  colmap::Reconstruction reconstruction;
-  std::vector<colmap::PosePrior> pose_priors;
+  auto reconstruction = std::make_shared<colmap::Reconstruction>();
 
-  InitializeGlomapFromDatabase(*database, reconstruction, view_graph);
+  GlobalMapper global_mapper(database);
+  global_mapper.BeginReconstruction(reconstruction);
 
-  GlobalMapper global_mapper(CreateTestOptions());
   std::unordered_map<frame_t, int> cluster_ids;
-  global_mapper.Solve(
-      database.get(), view_graph, reconstruction, pose_priors, cluster_ids);
+  global_mapper.Solve(CreateTestOptions(), cluster_ids);
 
   EXPECT_THAT(gt_reconstruction,
-              colmap::ReconstructionNear(reconstruction,
+              colmap::ReconstructionNear(*reconstruction,
                                          /*max_rotation_error_deg=*/1e-2,
                                          /*max_proj_center_error=*/1e-4));
 }
@@ -71,19 +65,16 @@ TEST(GlobalMapper, WithoutNoiseWithNonTrivialKnownRig) {
   colmap::SynthesizeDataset(
       synthetic_dataset_options, &gt_reconstruction, database.get());
 
-  ViewGraph view_graph;
-  colmap::Reconstruction reconstruction;
-  std::vector<colmap::PosePrior> pose_priors;
+  auto reconstruction = std::make_shared<colmap::Reconstruction>();
 
-  InitializeGlomapFromDatabase(*database, reconstruction, view_graph);
+  GlobalMapper global_mapper(database);
+  global_mapper.BeginReconstruction(reconstruction);
 
-  GlobalMapper global_mapper(CreateTestOptions());
   std::unordered_map<frame_t, int> cluster_ids;
-  global_mapper.Solve(
-      database.get(), view_graph, reconstruction, pose_priors, cluster_ids);
+  global_mapper.Solve(CreateTestOptions(), cluster_ids);
 
   EXPECT_THAT(gt_reconstruction,
-              colmap::ReconstructionNear(reconstruction,
+              colmap::ReconstructionNear(*reconstruction,
                                          /*max_rotation_error_deg=*/1e-2,
                                          /*max_proj_center_error=*/1e-4));
 }
@@ -105,28 +96,25 @@ TEST(GlobalMapper, WithoutNoiseWithNonTrivialUnknownRig) {
   colmap::SynthesizeDataset(
       synthetic_dataset_options, &gt_reconstruction, database.get());
 
-  ViewGraph view_graph;
-  colmap::Reconstruction reconstruction;
-  std::vector<colmap::PosePrior> pose_priors;
+  auto reconstruction = std::make_shared<colmap::Reconstruction>();
 
-  InitializeGlomapFromDatabase(*database, reconstruction, view_graph);
+  GlobalMapper global_mapper(database);
+  global_mapper.BeginReconstruction(reconstruction);
 
   // Set the rig sensors to be unknown
-  for (const auto& [rig_id, rig] : reconstruction.Rigs()) {
+  for (const auto& [rig_id, rig] : reconstruction->Rigs()) {
     for (const auto& [sensor_id, sensor] : rig.NonRefSensors()) {
       if (sensor.has_value()) {
-        reconstruction.Rig(rig_id).ResetSensorFromRig(sensor_id);
+        reconstruction->Rig(rig_id).ResetSensorFromRig(sensor_id);
       }
     }
   }
 
-  GlobalMapper global_mapper(CreateTestOptions());
   std::unordered_map<frame_t, int> cluster_ids;
-  global_mapper.Solve(
-      database.get(), view_graph, reconstruction, pose_priors, cluster_ids);
+  global_mapper.Solve(CreateTestOptions(), cluster_ids);
 
   EXPECT_THAT(gt_reconstruction,
-              colmap::ReconstructionNear(reconstruction,
+              colmap::ReconstructionNear(*reconstruction,
                                          /*max_rotation_error_deg=*/1e-2,
                                          /*max_proj_center_error=*/1e-4));
 }
@@ -149,19 +137,16 @@ TEST(GlobalMapper, WithNoiseAndOutliers) {
   colmap::SynthesizeNoise(
       synthetic_noise_options, &gt_reconstruction, database.get());
 
-  ViewGraph view_graph;
-  colmap::Reconstruction reconstruction;
-  std::vector<colmap::PosePrior> pose_priors;
+  auto reconstruction = std::make_shared<colmap::Reconstruction>();
 
-  InitializeGlomapFromDatabase(*database, reconstruction, view_graph);
+  GlobalMapper global_mapper(database);
+  global_mapper.BeginReconstruction(reconstruction);
 
-  GlobalMapper global_mapper(CreateTestOptions());
   std::unordered_map<frame_t, int> cluster_ids;
-  global_mapper.Solve(
-      database.get(), view_graph, reconstruction, pose_priors, cluster_ids);
+  global_mapper.Solve(CreateTestOptions(), cluster_ids);
 
   EXPECT_THAT(gt_reconstruction,
-              colmap::ReconstructionNear(reconstruction,
+              colmap::ReconstructionNear(*reconstruction,
                                          /*max_rotation_error_deg=*/1e-1,
                                          /*max_proj_center_error=*/1e-1,
                                          /*max_scale_error=*/std::nullopt,
