@@ -42,6 +42,9 @@ class TrackEngine {
       std::unordered_map<point3D_t, Point3D>& points3D_selected);
 
  private:
+  // Type alias for a 2D observation: (image_id, point2D_idx).
+  using Observation = std::pair<image_t, colmap::point2D_t>;
+
   // Blindly concatenate tracks if any matches occur
   void BlindConcatenation();
 
@@ -53,8 +56,8 @@ class TrackEngine {
   const ViewGraph& view_graph_;
   const std::unordered_map<image_t, colmap::Image>& images_;
 
-  // Internal structure used for concatenating tracks
-  colmap::UnionFind<image_pair_t> uf_;
+  // Internal structure used for concatenating tracks.
+  colmap::UnionFind<Observation> uf_;
 };
 
 }  // namespace glomap
