@@ -2,6 +2,12 @@
 
 namespace py = pybind11;
 
+void BindDepthMap(py::module& m);
+void BindNormalMap(py::module& m);
 void BindMVSModel(py::module& m);
 
-void BindMvs(py::module& m) { BindMVSModel(m); }
+void BindMvs(py::module& m) {
+  BindDepthMap(m);
+  BindNormalMap(m);
+  BindMVSModel(m);
+}
